Add nodemon in scripts and devDependencies
"nodemon server.js" will allow us to automatically restart the server whenever changes are made to the code. This is helpful during development as it saves us the manual effort of stopping and starting the server every time we make a change.

* Add nodemon in scripts and devDependencies in package.json

{
  "devDependencies": {
    "nodemon": "^3.0.1"
  },
  "scripts": {
    "dev": "nodemon server.js"
  }
}

In this example, we've defined a script named "dev" that runs our server.js file using nodemon.

* Then run "npm install" to install nodemon
* Then to start our Node.js server using "nodemon" we run "npm run dev"